move keyword cloud chart to a separate react component
