+++ client/src/assets/constants.js
@@ -1,0 +1,26 @@
+const colorMap = new Map([
+    [-10, "rgb(105, 212, 255)"],
+    [-9, "rgb(105, 212, 255)"],
+    [-8, "rgb(22, 141, 238)"],
+    [-7, "rgb(22, 141, 238)"],
+    [-6, "rgb(22, 69, 238)"],
+    [-5, "rgb(22, 69, 238)"],
+    [-4, "rgb(2, 40, 177)"],
+    [-3, "rgb(2, 40, 177)"],
+    [-2, "rgb(10,26,86)"],
+    [-1, "rgb(10,26,86)"],
+    [0, "rgb(72,22,122)"],
+    [1, "rgb(82,0,99)"],
+    [2, "rgb(82,0,99)"],
+    [3, "rgb(145, 0, 73)"],
+    [4, "rgb(145, 0, 73)"],
+    [5, "rgb(202, 57, 0)"],
+    [6, "rgb(202, 57, 0)"],
+    [7, "rgb(255,124,2)"],
+    [8, "rgb(255,124,2)"],
+    [9, "rgb(255,217,0)"],
+    [10, "rgb(255,217,0)"],
+])
+
+
+export { colorMap }
